SIXTH_PROJECT/requirements.txt
django=5.1.7
djangorestframework=3.15.2
psycopg2-binary=2.9.10
django-filter=25.1
python-dotenv=1.1.0

